+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+#### plt
+
+- `visible` member of SubplotDescriptor to control the visibility of an axis line
+
 ## [0.1.0] - 2022-08-30
 
 ### Added